+++ sora-android-sdk/src/main/kotlin/jp/shiguredo/sora/sdk/channel/rtc/PeerChannel.kt
@@ -118,6 +118,11 @@ interface PeerChannel {
             dataChannel: DataChannel,
         )
 
+        fun onDataChannelSignalingMessageSent(
+            label: String,
+            rawMessage: String,
+        ) {}
+
         fun onSenderEncodings(encodings: List<RtpParameters.Encoding>)
 
         fun onError(reason: SoraErrorReason)
@@ -658,7 +663,10 @@ class PeerChannelImpl(
         description: String,
     ) {
         val reAnswerMessage = MessageConverter.buildReAnswerMessage(description)
-        dataChannel.send(stringToDataChannelBuffer(dataChannel.label(), reAnswerMessage))
+        val sent = dataChannel.send(stringToDataChannelBuffer(dataChannel.label(), reAnswerMessage))
+        if (sent) {
+            listener?.onDataChannelSignalingMessageSent(dataChannel.label(), reAnswerMessage)
+        }
     }
 
     override fun sendStats(
@@ -677,7 +685,10 @@ class PeerChannelImpl(
         SoraLogger.d(TAG, "peer: sendDisconnectMessage, label=${dataChannel.label()}")
         val disconnectMessage = MessageConverter.buildDisconnectMessage(disconnectReason)
         SoraLogger.d(TAG, "peer: disconnectMessage=$disconnectMessage")
-        dataChannel.send(stringToDataChannelBuffer(dataChannel.label(), disconnectMessage))
+        val sent = dataChannel.send(stringToDataChannelBuffer(dataChannel.label(), disconnectMessage))
+        if (sent) {
+            listener?.onDataChannelSignalingMessageSent(dataChannel.label(), disconnectMessage)
+        }
     }
 
     override fun zipBufferIfNeeded(